cs493/hw4-chanall/requirements.txt
Flask==2.2.2
google-cloud-datastore==2.9.0